ENG-483 Create a Noop data fabric

+++ src/main/java/com/continuuity/performance/benchmark/Agent.java
@@ -2,7 +2,7 @@ package com.continuuity.performance.benchmark;
 
 public abstract class Agent {
 
-  public abstract void runOnce(int iteration, int agentId, int numAgents)
+  public abstract void runOnce(long iteration, int agentId, int numAgents)
       throws BenchmarkException;
 
 }
